Wrap WebSocket connection so we get nice encapsulation
